Corrected weight to double

+++ api/src/main/java/org/lushplugins/rewardsapi/api/reward/RandomReward.java
@@ -30,7 +30,7 @@ public class RandomReward extends Reward implements PlayerReward, LocationReward
 
         rewardMaps.forEach((rewardMap) -> {
             Reward reward = Reward.loadReward(rewardMap, rewardMap.toString());
-            int weight = rewardMap.containsKey("weight") ? (int) rewardMap.get("weight") : 1;
+            double weight = rewardMap.containsKey("weight") ? (double) rewardMap.get("weight") : 1;
             if (reward != null) {
                 rewards.add(reward, weight);
             }
